Struct led_cube::Cube

pub struct Cube {
    // some fields omitted
}

A connection to an LED cube.

The methods on this struct operate on an internal buffer, and won't have an effect on the cube until the flush method is called.

Examples

Turn each light on in sequence:

use std::thread;
use std::time::Duration;
use led_cube::*;

let mut cube = Cube::new("COM5").unwrap();

for z in 0..4 {
    for y in 0..4 {
        for x in 0..4 {
            cube.clear();
            cube.set([x, y, z], true);
            cube.flush();
            thread::sleep(Duration::from_millis(500));
        }
    }
}

Methods

impl Cube

fn new<T: AsRef<OsStr> + ?Sized>(port: &T) -> Result<Self>

Connects to an LED cube through a serial port.

fn flush(&mut self) -> Result<()>

Updates the LED cube to match the internal buffer.

fn set(&mut self, pos: CubePosition, state: bool)

Turns the LED at position pos on or off based on state.

Panics

Panics if pos is out of range.

fn get(&self, pos: CubePosition) -> bool

Returns true if the LED at position pos is on.

Panics

Panics if pos is out of range.

fn clear(&mut self)

Turns off all LEDs.